+++ src/LazySeoToolsServiceProvider.php
@@ -5,4 +5,6 @@ namespace Step2dev\LazySeoTools;
 /**
  * @deprecated Use LazySeoServiceProvider. Kept for backward compatibility.
  */
-class LazySeoToolsServiceProvider extends LazySeoServiceProvider {}
+class LazySeoToolsServiceProvider extends LazySeoServiceProvider
+{
+}